% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_unique.R
\name{all_unique}
\alias{all_unique}
\title{Test that all values are unique in an array or vector}
\usage{
all_unique(x)
}
\arguments{
\item{x}{A numeric or array object.}
}
\value{
logical
}
\description{
Test that all values are unique in an array or vector
}
