% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticize.R
\name{attempt_latticization}
\alias{attempt_latticization}
\title{Attempt to swap the target vertices between a pair of edges}
\usage{
attempt_latticization(x, d, allow_disconnect)
}
\arguments{
\item{x}{An adjacency matrix or igraph object representing the graph to
mutate.}

\item{d}{A square matrix with as many rows and columns as vertices in
\code{x}, which expresses distance to the diagonal. A cannonical \code{d}
can be generated with the \code{\link{distance_to_diagonal}} function.}

\item{allow_disconnect}{A logical indicating where a mutation is allowed to
disrupt the connectedness of the network. If \code{FALSE}, then \code{x}
must be connected to begin with and represented as an \code{igraph} object.
The connectedness check is not implemented for plain adjacency matrices.}
}
\value{
A mutated version of \code{x}.
}
\description{
This will typically be used in a loop to incrementally mutate a graph to
become more lattice-like.
}
\details{
If \code{allow_disconnect == TRUE}, the procedure does not require
  \code{igraph}. If \code{FALSE}, \code{igraph} is required.

  A mutation will occur only under the following conditions:
  1. The mutation will create new edges.
  2. The mutation will create edges that correspond to smaller values in
  \code{d}.
  3. If \code{allow_disconnect == FALSE}, the mutation must not disrupt
  network connectedness.

  The test for connectedness is slow compared to the other tests and the
  rewiring operation itself.
}
