% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_diagonal.R
\name{distance_to_diagonal}
\alias{distance_to_diagonal}
\title{Generate matrix of distances from its own diagonal}
\usage{
distance_to_diagonal(n)
}
\arguments{
\item{n}{Number of vertices}
}
\value{
Matrix
}
\description{
Generate matrix of distances from its own diagonal
}
\details{
Distances are determined assuming that the list of vertices loops
  back to the start. If there are 4 vertices, the output will be:
  0 1 2 1
  1 0 1 2
  2 1 0 1
  1 2 1 0
  The value at row 1 col 4 is 1 because vertex 4 is considered to be adjacent
  to vertex 1.
}
