% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge2vertices.R, R/vertices2edge.R
\name{edge2vertices}
\alias{edge2vertices}
\alias{edge2vertices.matrix}
\alias{edge2vertices.igraph}
\alias{vertices2edge}
\alias{vertices2edge.matrix}
\alias{vertices2edge.igraph}
\title{Translate between index and vertex edge references}
\usage{
edge2vertices(x, ...)

\method{edge2vertices}{matrix}(x, indices)

\method{edge2vertices}{igraph}(g, indices)

vertices2edge(x, ...)

\method{vertices2edge}{matrix}(x, vertices)

\method{vertices2edge}{igraph}(g, vertices)
}
\arguments{
\item{x}{An adjacency matrix}

\item{indices}{A numeric vector; 1-D index referencing edges in the graph.}

\item{g}{An igraph object}

\item{vertices}{A matrix specifying edges as source->target ordered pairs.}
}
\value{
A matrix of vertex pairs representing edges. Each column is an edge,
  and the first row is the source vertex.
}
\description{
Vertices refer to the \code{[source, target]} ordered pair defining a
directed edge. Indices refer to a 1-D index into the graph representation.
The functions \code{\link{edge2vertices}} and \code{\link{vertices2edge}}
translate between these representations.
}
\details{
Adjacency matrices are assumed to be oriented such that the row
  indicates the source vertex and the column indicates that target vertex.

  The returned matrix of \code{vertices} will be oriented such that each
  column is an edge. The first row contains the source vertices, and the
  second row contains the target vertices.
}
