% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_exists.R
\name{edge_exists}
\alias{edge_exists}
\title{Check that edges exist in a graph}
\usage{
edge_exists(x, ...)
}
\arguments{
\item{x}{An adjacency matrix.}

\item{g}{An igraph object.}

\item{edges}{A matrix of source->target ordered pairs representing edges.}
}
\value{
logical.
}
\description{
Edges are specified by directed source->target pairs arranged in a matrix.
}
\details{
Even if querying a single edge, the input must be represented as a
  matrix. In the case of one edge, the matrix would have two rows and one
  column.
}
