% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_indices.R
\name{edge_indices}
\alias{edge_indices}
\alias{edge_indices.matrix}
\alias{edge_indices.igraph}
\title{Return edge indices}
\usage{
edge_indices(x)

\method{edge_indices}{matrix}(x)

\method{edge_indices}{igraph}(g)
}
\arguments{
\item{x, g}{An adjacency matrix or igraph object.}
}
\value{
Indices corresponding to existing edges.
}
\description{
Return edge indices
}
\details{
If the input is an adjacency matrix, the return value is a 1-D index
  identifying the non-zero values in the matrix. If the input is an igraph,
  then the return value is an \code{igraph.es} object---the output of
  \code{\link[igraph]{E}}.
}
