% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticize.R
\name{latticize}
\alias{latticize}
\title{Lattice with preserved in/out degree distribution}
\usage{
latticize(x, epochs, allow_disconnect = TRUE)
}
\description{
This function "latticizes" a directed network, while preserving the in-
  and out-degree distributions. In weighted networks, the function
  preserves the out-strength but not the in-strength distributions. The
  function can ensure that the randomized network maintains
  connectedness, the ability for every node to reach every other node in
  the network. If this is desired, the input network for this function must
  be connected.
}
\details{
@param x An adjacency matrix or igraph object representing the graph to
    mutate. It is assumed to be directed and weighted, but unweighted is
    fine.
  @param epochs rewiring parameter
              (each edge is rewired approximately ITER times)
  @param allow_disconnect A logical indicating where a mutation is allowed to
    disrupt the connectedness of the network. If \code{FALSE}, then \code{x}
    must be connected to begin with and represented as an \code{igraph} object.
    The connectedness check is not implemented for plain adjacency matrices.

  @return A mutated version of the input network.

  @references Maslov and Sneppen (2002) Science 296:910
              Sporns and Zwi (2004) Neuroinformatics 2:145

  Mika Rubinov, UNSW, 2007-2010
  Olaf Sporns, Indiana University, 2012
  Chris Cox, Louisiana State University, 2021 (R port)
}
