% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticize.R
\name{max_attempts}
\alias{max_attempts}
\title{Equation for setting the maximum number of attempts}
\usage{
max_attempts(v, e)
}
\arguments{
\item{v}{number of vertices}

\item{e}{number of edges}
}
\value{
An integer
}
\description{
Equation for setting the maximum number of attempts
}
