% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_edges.R
\name{n_edges}
\alias{n_edges}
\alias{n_edges.matrix}
\alias{n_edges.igraph}
\title{Count the number of edges in the graph}
\usage{
n_edges(x, ...)

\method{n_edges}{matrix}(x, count_loops = TRUE)

\method{n_edges}{igraph}(g, count_loops = TRUE)
}
\arguments{
\item{x}{An adjacency matrix}

\item{count_loops}{Logical. Should loops (the diagonal) be included in the
count?}

\item{g}{An igraph object.}
}
\value{
Number of edges
}
\description{
A common interface for matrix and igraph representations of graph structure.
}
