% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_vertices.R
\name{n_vertices}
\alias{n_vertices}
\alias{n_vertices.matrix}
\alias{n_vertices.igraph}
\title{Count the number of edges in the graph}
\usage{
n_vertices(x)

\method{n_vertices}{matrix}(x)

\method{n_vertices}{igraph}(g)
}
\arguments{
\item{x}{An adjacency matrix}

\item{g}{An igraph object.}
}
\value{
Number of edges
}
\description{
A common interface for matrix and igraph representations of graph structure.
}
