% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewire_targets.R
\name{rewire_targets}
\alias{rewire_targets}
\alias{rewire_targets.matrix}
\alias{rewire_targets.igraph}
\title{Mutate the graph to swap targets between a pair of edges}
\usage{
rewire_targets(x, ...)

\method{rewire_targets}{matrix}(x, edges)

\method{rewire_targets}{igraph}(g, edges)
}
\arguments{
\item{x}{An adjacency matrix}

\item{edges}{A matrix where each column is a source->destination ordered pair
representing an edge.}

\item{g}{An igraph object}
}
\value{
A mutated version of the graph.
}
\description{
Mutate the graph to swap targets between a pair of edges
}
