% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_creates_new_edges.R
\name{rewiring_creates_new_edges}
\alias{rewiring_creates_new_edges}
\title{Check that swapping targets between two edges will require creating new edges}
\usage{
rewiring_creates_new_edges(x, edges)
}
\arguments{
\item{x}{An adjacency matrix or igraph object}

\item{edges}{A matrix specifying edges as source->destination ordered
pairs.}
}
\value{
logical
}
\description{
Check that swapping targets between two edges will require creating new edges
}
\details{
\code{edges} should be oriented such that each column is an
  edge, the first row contains the source vertices, and the second row
  contains the target vertices.
}
