% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_maintains_connectedness.R
\name{rewiring_maintains_connectedness}
\alias{rewiring_maintains_connectedness}
\alias{rewiring_maintains_connectedness.igraph}
\title{Check that swapping targets between two edges maintains conncectedness}
\usage{
rewiring_maintains_connectedness(x, ...)

\method{rewiring_maintains_connectedness}{igraph}(g, edges)
}
\arguments{
\item{x, g}{An adjacency matrix or igraph object}

\item{edges}{A matrix where each column is a source->destination ordered pair
representing an edge.}
}
\value{
logical
}
\description{
Connectedness refers to the ability to get between any pair of vertices in
the graph.
}
