% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_makes_more_lattice_like.R
\name{rewiring_makes_more_lattice_like}
\alias{rewiring_makes_more_lattice_like}
\title{Check that mutating the graph make it more like a lattice}
\usage{
rewiring_makes_more_lattice_like(x, d, edges)
}
\arguments{
\item{x}{An adjacency matrix or igraph object}

\item{d}{A matrix specifying the "cost" of each edge}

\item{edges}{A matrix specifying edges as source->target ordered pairs.}
}
\value{
logical
}
\description{
The objective is to reduce the overall cost of edges with respect to
\code{d}. If \code{d} is defined so that cost increases with distance from
the diagonal, reducing cost will encourage a lattice-like structure.
}
