% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_vertices.R
\name{shuffle_vertices}
\alias{shuffle_vertices}
\alias{shuffle_vertices.matrix}
\alias{shuffle_vertices.igraph}
\alias{unshuffle_vertices}
\alias{unshuffle_vertices.matrix}
\alias{unshuffle_vertices.igraph}
\title{Shuffle the vertices of a graph or adjacency matrix}
\usage{
shuffle_vertices(x)

\method{shuffle_vertices}{matrix}(x)

\method{shuffle_vertices}{igraph}(g)

unshuffle_vertices(x)

\method{unshuffle_vertices}{matrix}(x)

\method{unshuffle_vertices}{igraph}(g)
}
\arguments{
\item{x}{An adjacency \code{matrix}}

\item{g}{An \code{igraph} object.}
}
\value{
A shuffled version of the input
}
\description{
Vertices are randomly sampled without replacement to generate a shuffled
version of the input.
}
\details{
The index used to shuffle the vertices is stored as an attribute,
  which is referenced by the \code{unshuffle_vertices} function to get back
  to the original order.
}
