% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_targets.R
\name{swap_targets}
\alias{swap_targets}
\title{Swap targets between a pair of edges}
\usage{
swap_targets(edges)
}
\arguments{
\item{edges}{A 2x2 matrix. Each column is an edge defined by a source
(row 1) and target (row 2) vertex.}
}
\value{
A matrix with the same dimensions as \code{edges}, but with
}
\description{
Given two directed edges a->b and c->d, return a->d and c->b.
}
